+++ webserver/tests/test_tasks.py
@@ -1,12 +1,13 @@
 import json
 import pytest
+from copy import deepcopy
 from unittest import mock
 from datetime import datetime, timedelta
 from kubernetes.client.exceptions import ApiException
 from unittest import mock
 from unittest.mock import Mock
 
-from app.helpers.const import CLEANUP_AFTER_DAYS
+from app.helpers.const import CLEANUP_AFTER_DAYS, TASK_POD_RESULTS_PATH
 from app.helpers.db import db
 from app.helpers.exceptions import InvalidRequest
 from app.models.task import Task
@@ -14,7 +15,7 @@ from app.models.task import Task
 
 @pytest.fixture(scope='function')
 def task_body(dataset, image_name):
-    return {
+    return deepcopy({
         "name": "Test Task",
         "requested_by": "das9908-as098080c-9a80s9",
         "executors": [
@@ -35,8 +36,8 @@ def task_body(dataset, image_name):
         "inputs":{},
         "outputs":{},
         "resources": {},
-        "volumes": {},
-    }
+        "volumes": {}
+    })
 
 @pytest.fixture
 def running_state():
@@ -105,7 +106,7 @@ def test_get_list_tasks_base_user(
 
 def test_create_task(
         cr_client,
-        k8s_client_task,
+        k8s_client,
         post_json_admin_header,
         client,
         task_body
@@ -113,18 +114,59 @@ def test_create_task(
     """
     Tests task creation returns 201
     """
-    data = task_body
-
     response = client.post(
         '/tasks/',
-        data=json.dumps(data),
+        data=json.dumps(task_body),
         headers=post_json_admin_header
     )
     assert response.status_code == 201
 
+def test_create_task_invalid_output_field(
+        cr_client,
+        k8s_client,
+        post_json_admin_header,
+        client,
+        task_body
+    ):
+    """
+    Tests task creation returns 4xx request when output
+    is not a dictionary
+    """
+    task_body["outputs"] = []
+    response = client.post(
+        '/tasks/',
+        data=json.dumps(task_body),
+        headers=post_json_admin_header
+    )
+    assert response.status_code == 400
+    assert response.json == {"error": "\"outputs\" filed muct be a json object or dictionary"}
+
+def test_create_task_no_output_field_reverts_to_default(
+        cr_client,
+        k8s_client,
+        post_json_admin_header,
+        client,
+        task_body
+    ):
+    """
+    Tests task creation returns 201 but the volume mounted
+    is the default one
+    """
+    task_body.pop("outputs")
+    response = client.post(
+        '/tasks/',
+        data=json.dumps(task_body),
+        headers=post_json_admin_header
+    )
+    assert response.status_code == 201
+    k8s_client["create_namespaced_pod_mock"].assert_called()
+    pod_body = k8s_client["create_namespaced_pod_mock"].call_args.kwargs["body"]
+    assert len(pod_body.spec.containers[0].volume_mounts) == 1
+    assert pod_body.spec.containers[0].volume_mounts[0].mount_path == TASK_POD_RESULTS_PATH
+
 def test_create_task_with_ds_name(
         cr_client,
-        k8s_client_task,
+        k8s_client,
         post_json_admin_header,
         client,
         dataset,
@@ -146,7 +188,7 @@ def test_create_task_with_ds_name(
 
 def test_create_task_with_ds_name_and_id(
         cr_client,
-        k8s_client_task,
+        k8s_client,
         post_json_admin_header,
         client,
         dataset,
@@ -167,7 +209,7 @@ def test_create_task_with_ds_name_and_id(
 
 def test_create_task_with_conflicting_ds_name_and_id(
         cr_client,
-        k8s_client_task,
+        k8s_client,
         post_json_admin_header,
         client,
         dataset,
@@ -275,7 +317,7 @@ def test_create_task_image_not_found(
 def test_get_task_by_id_admin(
         token_valid_mock,
         cr_client,
-        k8s_client_task,
+        k8s_client,
         post_json_admin_header,
         post_json_user_header,
         simple_admin_header,
@@ -304,7 +346,7 @@ def test_get_task_by_id_admin(
 def test_get_task_by_id_non_admin_owner(
         token_valid_mock,
         cr_client,
-        k8s_client_task,
+        k8s_client,
         simple_user_header,
         post_json_user_header,
         client,
@@ -331,8 +373,7 @@ def test_get_task_by_id_non_admin_owner(
 def test_get_task_by_id_non_admin_non_owner(
         token_valid_mock,
         cr_client,
-        k8s_client_task,
-        access_request,
+        k8s_client,
         post_json_user_header,
         simple_user_header,
         client,
@@ -361,7 +402,7 @@ def test_get_task_by_id_non_admin_non_owner(
 def test_cancel_task(
         client,
         cr_client,
-        k8s_client_task,
+        k8s_client,
         simple_admin_header,
         post_json_admin_header,
         task_body
@@ -474,7 +515,7 @@ class TestTaskResults:
         client,
         task_body,
         mocker,
-        mock_k8s_client
+        k8s_client
     ):
         """
         A simple test with mocked PVs to test a successful result
@@ -484,10 +525,6 @@ class TestTaskResults:
         data = task_body
         # The mock has to be done manually rather than use the fixture
         # as it complains about the return value of the list_pod method
-        mocker.patch(
-            'app.models.task.KubernetesClient',
-            return_value=mock_k8s_client
-        )
         mocker.patch('app.models.task.uuid4', return_value="1dc6c6d1-417f-409a-8f85-cb9d20f7c741")
         response = client.post(
             '/tasks/',
@@ -496,20 +533,12 @@ class TestTaskResults:
         )
         assert response.status_code == 201
 
-        # Get results
-        mocker.patch(
-            'app.models.task.KubernetesBatchClient'
-        )
-        k8s_client = mocker.patch(
-            'app.models.task.KubernetesClient',
-            return_value=Mock(list_namespaced_pod=Mock())
-        )
         pod_mock = Mock()
         pod_mock.metadata.labels = {"job-name": "result-job-1dc6c6d1-417f-409a-8f85-cb9d20f7c741"}
         pod_mock.metadata.name = "result-job-1dc6c6d1-417f-409a-8f85-cb9d20f7c741"
         pod_mock.spec.containers = [Mock(image=task_body["executors"][0]["image"])]
         pod_mock.status.container_statuses = [Mock(ready=True)]
-        k8s_client.return_value.list_namespaced_pod.return_value.items = [pod_mock]
+        k8s_client["list_namespaced_pod_mock"].return_value.items = [pod_mock]
 
         mocker.patch(
             'app.models.task.Task.get_status',
@@ -531,7 +560,7 @@ class TestTaskResults:
         client,
         task_body,
         mocker,
-        mock_k8s_client
+        k8s_client
     ):
         """
         Tests that the job creation to fetch results from a PV returns a 500
@@ -539,12 +568,7 @@ class TestTaskResults:
         """
         # Create a new task
         data = task_body
-        # The mock has to be done manually rather than use the fixture
-        # as it complains about the return value of the list_pod method
-        mocker.patch(
-            'app.models.task.KubernetesClient',
-            return_value=mock_k8s_client
-        )
+
         response = client.post(
             '/tasks/',
             data=json.dumps(data),
@@ -553,19 +577,14 @@ class TestTaskResults:
         assert response.status_code == 201
 
         # Get results - creating a job fails
-        k8s_batch = mocker.patch('app.models.task.KubernetesBatchClient')
-        k8s_batch.return_value.create_namespaced_job.side_effect = ApiException(status=500, reason="Something went wrong")
+        k8s_client["create_namespaced_job_mock"].side_effect = ApiException(status=500, reason="Something went wrong")
 
-        k8s_client = mocker.patch(
-            'app.models.task.KubernetesClient',
-            return_value=Mock(list_namespaced_pod=Mock())
-        )
         pod_mock = Mock()
         pod_mock.metadata.labels = {"job-name": "result-job-1dc6c6d1-417f-409a-8f85-cb9d20f7c741"}
         pod_mock.metadata.name = "result-job-1dc6c6d1-417f-409a-8f85-cb9d20f7c741"
         pod_mock.spec.containers = [Mock(image=task_body["executors"][0]["image"])]
         pod_mock.status.container_statuses = [Mock(ready=True)]
-        k8s_client.return_value.list_namespaced_pod.return_value.items = [pod_mock]
+        k8s_client["list_namespaced_pod_mock"].return_value.items = [pod_mock]
 
         response = client.get(
             f'/tasks/{response.json["task_id"]}/results',
@@ -602,7 +621,7 @@ class TestTaskResults:
 
 def test_get_task_status_running_and_waiting(
     cr_client,
-    k8s_client_task,
+    k8s_client,
     running_state,
     waiting_state,
     post_json_admin_header,
@@ -657,7 +676,7 @@ def test_get_task_status_running_and_waiting(
 
 def test_get_task_status_terminated(
     cr_client,
-    k8s_client_task,
+    k8s_client,
     terminated_state,
     post_json_admin_header,
     client,