+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Releases Changelog
 
 ## 0.5.2
+- POST `/tasks` now accepts the outputs field to dynamically mount a volume so that results can be fetched correctly. If no value is provided, the default location of `/mnt/data/` will be used.
 
 - Added PATCH /datasets/<id> endpoints, so existing datasets can be amended, or a dictionary added to them.